Refactor: 입력검증기능_예외처리 수정
