+++ src/main/java/lotto/Application.java
@@ -1,7 +1,5 @@
 package lotto;
 
-import java.util.List;
-
 public class Application {
     private static Run run = new Run();
