+++ src/main/java/lotto/view/ViewInput.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
+import lotto.verification.InputVerification;
 
 import java.util.ArrayList;
 import java.util.List;
