+++ src/test/java/lotto/GameTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.Game;
+import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
